Rename webpack configs
