+++ client/package.json
@@ -7,9 +7,9 @@
     "npm": "4.1.1"
   },
   "scripts": {
-    "build:test": "webpack --config webpack.config.js",
-    "build:production": "NODE_ENV=production webpack --config webpack.config.js",
-    "build:development": "NODE_ENV=development webpack -w --config webpack/development.js",
+    "build:test": "webpack --config --config webpack/webpack.config.test.js",
+    "build:production": "NODE_ENV=production webpack --config webpack/webpack.config.production.js",
+    "build:development": "NODE_ENV=development webpack -w --config webpack/webpack.config.development.js",
     "lint": "eslint --ext .js,.jsx ."
   },
   "cacheDirectories": ["node_modules", "client/node_modules"],
