+++ client/webpack/loaders/babel.js
@@ -1,10 +1,5 @@
 module.exports = {
   test: /\.(js|jsx)?(\.erb)?$/,
   exclude: /node_modules/,
-  loader: 'babel-loader',
-  options: {
-    presets: [
-      'es2015', 'stage-2', 'react',
-    ],
-  },
+  use: 'babel-loader',
 };
