+++ client/webpack/webpack.config.development.server.js
@@ -1,9 +1,7 @@
-// Note: You must restart bin/webpack-dev-server for changes to take effect
-
 const { resolve } = require('path');
 const merge = require('webpack-merge');
-const devConfig = require('./development.js');
-const { devServer, publicPath, paths } = require('./configuration.js');
+const devConfig = require('./webpack.config.development.js');
+const { devServer, publicPath, paths } = require('./webpackConfigLoader.js');
 
 module.exports = merge(devConfig, {
   devServer: {
