+++ client/webpack/webpack.config.production.js
@@ -1,10 +1,9 @@
 /* eslint global-require: 0 */
-// Note: You must run bin/webpack for changes to take effect
 
 const webpack = require('webpack');
 const merge = require('webpack-merge');
 const CompressionPlugin = require('compression-webpack-plugin');
-const sharedConfig = require('./shared.js');
+const sharedConfig = require('./webpack.config.base.js');
 
 module.exports = merge(sharedConfig, {
   output: { filename: '[name]-[chunkhash].js' },
