+++ client/webpack/webpack.config.test.js
@@ -1,0 +1,4 @@
+const merge = require('webpack-merge');
+const sharedConfig = require('./webpack.config.base.js');
+
+module.exports = merge(sharedConfig, {});
