+++ config/webpack/paths.yml
@@ -1,7 +1,11 @@
 # Restart webpack-watcher or webpack-dev-server if you make changes here
 default: &default
   config: client/webpack
-  entry: HelloWorld/startup/registration
+  entry:
+    - HelloWorld/startup/registration
+    - es5-shim/es5-shim
+    - es5-shim/es5-sham
+    - babel-polyfill
   output: public/assets/webpack
   manifest: manifest.json
   node_modules: client/node_modules
